Chore: 찜하기 코드 수정

+++ src/main/java/com/palleteforco/palleteforco/domain/dib/service/DibServiceImpl.java
@@ -2,10 +2,11 @@ package com.palleteforco.palleteforco.domain.dib.service;
 
 import com.palleteforco.palleteforco.domain.dib.dto.DibDto;
 import com.palleteforco.palleteforco.domain.dib.mapper.DibMapper;
+import com.palleteforco.palleteforco.global.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class DibServiceImpl implements DibService{
@@ -17,14 +18,21 @@ public class DibServiceImpl implements DibService{
     }
 
     public void addDib(DibDto dibDto) throws Exception {
+        OAuth2User oAuth2User = (OAuth2User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = (String)oAuth2User.getAttributes().get("email");
+
+        dibDto.setEmail(email);
+
         dibMapper.insertDib(dibDto);
     }
 
-    public List<DibDto> getDibList() throws Exception {
-        return dibMapper.selectDibList();
-    }
-
     public void cancelDib(Long dib_id) throws Exception {
+        DibDto existing = dibMapper.selectDib(dib_id);
+
+        if (existing == null) {
+            throw new NotFoundException("좋아요 한 게시물이 없습니다.");
+        }
+
         dibMapper.deleteDib(dib_id);
     }
 }
